test(create-purista): add parseArgs coverage
