+++ src/index.ts
@@ -3,40 +3,7 @@ import { createTerminalPromptAdapter } from '../../purista/packages/cli/src/adap
 import { createTerminalOutputAdapter } from '../../purista/packages/cli/src/adapters/output/terminalOutput.js'
 import { PuristaCliError } from '../../purista/packages/cli/src/core/errors.js'
 import { createPuristaCliEngine } from '../../purista/packages/cli/src/engine.js'
-
-type ParsedArgs = {
-	_: string[]
-	[key: string]: string | boolean | string[] | undefined
-}
-
-const parseArgs = (argv: string[]): ParsedArgs => {
-	const result: ParsedArgs = { _: [] }
-
-	for (let index = 0; index < argv.length; index += 1) {
-		const token = argv[index]
-		if (!token.startsWith('--')) {
-			result._.push(token)
-			continue
-		}
-
-		const normalized = token.slice(2)
-		if (normalized.startsWith('no-')) {
-			result[normalized.slice(3)] = false
-			continue
-		}
-
-		const next = argv[index + 1]
-		if (!next || next.startsWith('--')) {
-			result[normalized] = true
-			continue
-		}
-
-		result[normalized] = next
-		index += 1
-	}
-
-	return result
-}
+import { parseArgs } from './parseArgs.js'
 
 async function main() {
 	const args = parseArgs(process.argv.slice(2))
